# api-doc-swagger.yml
openapi: 3.0.0
info:
  description: 'this is API for Orion Store Ecommerce. There are public, authentificated and authorized APIs. <ul> <ol> For Auth APIs  go to [https://orion-api-five.vercel.app/api/public/tokens?template=user_token](https://orion-api-five.vercel.app/api/public/tokens?template=user_token) </ol> <ol> To get authorized token go to  [https://orion-api-five.vercel.app/api/public/tokens?template=admin_token](https://orion-api-five.vercel.app/api/public/tokens?template=admin_token) </ol>'
  version: 1.0.0
  title: API for Orion Store Multi Vendor SASS.
  contact:
    name: Sylvain Codes
    url: 'https://www.patreon.com/sylvaincodes'
    email: sylvaincodeur@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  termsOfService: 'https://www.patreon.com/sylvaincodes'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: 'https://orion-api-five.vercel.app/api'
...
paths:
...
  /user/addresses:
    get:
      tags:
        - addresses
      summary: Find all addresses published by customer
      parameters:
        - name: user_id
          in: query
          description: this is the id of user
          required: true
          schema:
            type: string
      description: This api is protected by clerk auth.
      security:
        - bearerAuth:
            - 'read:addresses'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
    post:
      tags:
        - addresses
      summary: Create a new addresses
      description: This api is protected by clerk auth.
      requestBody:
        $ref: '#/components/requestBodies/Address'
      security:
        - bearerAuth:
            - 'write:addresses'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
    delete:
      tags:
        - addresses
      summary: Delete a addresses
      description: This api is protected by clerk auth.
      parameters:
        - name: _id
          in: query
          description: delete an address
          required: true
          schema:
            type: string
      security:
        - bearerAuth:
            - 'write:addresses'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /user/brands:
    get:
      tags:
        - brands
      summary: Find one brand  published by admin
      description: This api is protected by auth.
      parameters:
        - name: _id
          in: query
          description: this is the id of brand
          required: false
          schema:
            type: string
      security:
        - bearerAuth:
            - 'read:brands'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
components:
  schemas:
...
    Address:
      type: object
      required:
        - firstName
        - lastName
        - address
        - city
        - state
        - phone
        - country
        - zipCode
        - email
        - user_id
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        country:
          type: string
        zipCode:
          type: string
        email:
          type: string
        user_id:
          type: string
        status:
          type: string
          enum:
            - draft
            - publish
            - archive
          default: publish
      description: Address model
...
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: number
          enum:
            - 401
          default: 401
    Response:
      type: object
      required:
        - status
        - data
      properties:
        message:
          type: string
        status:
          type: number
          default: 200
        success:
          type: boolean
          default: true
        data:
          type: array
          items:
            type: object
    PostResponse:
      type: object
      required:
        - status
        - data
      properties:
        message:
          type: string
        status:
          type: number
          default: 200
        success:
          type: boolean
          default: true
        data:
          type: object
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: number
          default: 204
        success:
          type: boolean
          default: true
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
  requestBodies:
    Address:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Address'
      description: Address object that needs to be added to the form
      required: true